Add requirements and renamed the file
This commit adds the requirement and rename the file for proper convention,
this also introduces the changes to make the location taken by user as well as
passed by the command line arguement.
