+++ requirements.txt
@@ -3,3 +3,4 @@ numpy==1.11.0
 matplotlib==2.2.2
 python_dateutil==2.7.2
 tabulate==0.8.2
+pandas==0.23.1
